+++ src/luaopen/utility_functions.cpp
@@ -23,8 +23,11 @@
 #include "../utils/VariantArguments.hpp"
 #include "../utils/function_wrapper.hpp"
 #include "../utils/string_literal.hpp"
+#include "../LuaCoroutine.hpp"
+#include "../LuaObject.hpp"
 
 #include <godot_cpp/core/error_macros.hpp>
+#include <godot_cpp/classes/object.hpp>
 #include <godot_cpp/variant/utility_functions.hpp>
 
 using namespace godot;
@@ -45,6 +48,71 @@ template<typename RetType, StringLiteral func_name, size_t func_hash> sol::objec
 	}
 }
 
+struct ResumeLuaCoroutineCallable : public CallableCustom {
+	ResumeLuaCoroutineCallable(lua_State *L)
+		: coroutine(LuaObject::wrap_object<LuaCoroutine>(sol::thread(L, L)))
+	{
+	}
+
+	uint32_t hash() const override {
+		return get_as_text().hash();
+	}
+
+	String get_as_text() const override {
+		return "<ResumeLuaCoroutineCallable>";
+	}
+
+	CompareEqualFunc get_compare_equal_func() const override {
+		return nullptr;
+	}
+
+	CompareLessFunc get_compare_less_func() const override {
+		return nullptr;
+	}
+
+	bool is_valid() const override {
+		return coroutine.is_valid()
+			&& (coroutine->get_status() == LuaCoroutine::STATUS_OK || coroutine->get_status() == LuaCoroutine::STATUS_YIELD);
+	}
+	
+	ObjectID get_object() const override {
+		return {};
+	}
+
+	void call(const Variant **p_arguments, int p_argcount, Variant &r_return_value, GDExtensionCallError &r_call_error) const override {
+		r_return_value = coroutine->resume(p_arguments, p_argcount, r_call_error);
+	}
+
+	Ref<LuaCoroutine> coroutine;
+};
+
+static void lua_await(sol::this_state L, sol::stack_object signal_or_coroutine) {
+	ERR_FAIL_COND_MSG(!lua_isyieldable(L), "Current thread is not yieldable");
+	
+	Signal signal;
+	Variant var = to_variant(signal_or_coroutine);
+	switch (var.get_type()) {
+		case Variant::Type::SIGNAL:
+			signal = var;
+			break;
+
+		case Variant::Type::OBJECT:
+			if (Object *obj = var; obj->has_signal("completed")) {
+				signal = Signal(obj, "completed");
+			}
+			break;
+		
+		default:
+			break;
+	}
+	
+	ERR_FAIL_COND_MSG(signal.is_null(), "Expected signal in await");
+	
+	Callable callback(memnew(ResumeLuaCoroutineCallable(L)));
+	signal.connect(callback, Object::CONNECT_ONE_SHOT);
+	lua_yield(L, 0);
+}
+
 extern "C" int luaopen_godot_utility_functions(lua_State *L) {
 	sol::state_view state = L;
 
@@ -52,6 +120,9 @@ extern "C" int luaopen_godot_utility_functions(lua_State *L) {
 
 	// In Lua, `print` separates passed values with "\t", so we bind it to Godot's `printt`
 	state.do_string("print = printt");
+
+	// Use `await` to await for signals.
+	state.set("await", lua_await);
 
 	return 0;
 }