+++ src/script-language/LuaScriptInstance.cpp
@@ -27,9 +27,11 @@
 #include "LuaScriptLanguage.hpp"
 #include "LuaScriptMetadata.hpp"
 #include "LuaScriptProperty.hpp"
+#include "../LuaCoroutine.hpp"
 #include "../LuaError.hpp"
 #include "../LuaFunction.hpp"
 #include "../LuaTable.hpp"
+#include "../utils/VariantArguments.hpp"
 #include "../utils/convert_godot_lua.hpp"
 
 namespace luagdextension {
@@ -54,7 +56,7 @@ LuaScriptInstance::~LuaScriptInstance() {
 GDExtensionBool set_func(LuaScriptInstance *p_instance, const StringName *p_name, const Variant *p_value) {
 	// 1) try calling `_set`
 	if (const LuaScriptMethod *_set = p_instance->script->get_metadata().methods.getptr("_set")) {
-		Variant value_was_set = LuaFunction::invokev_lua(_set->method, Array::make(p_instance->owner, *p_name, *p_value), false);
+		Variant value_was_set = LuaCoroutine::invoke_lua(_set->method, Array::make(p_instance->owner, *p_name, *p_value), false);
 		if (value_was_set) {
 			return true;
 		}
@@ -79,7 +81,7 @@ GDExtensionBool set_func(LuaScriptInstance *p_instance, const StringName *p_name
 GDExtensionBool get_func(LuaScriptInstance *p_instance, const StringName *p_name, Variant *p_value) {
 	// a) try calling `_get`
 	if (const LuaScriptMethod *_get = p_instance->script->get_metadata().methods.getptr("_get")) {
-		Variant value = LuaFunction::invokev_lua(_get->method, Array::make(p_instance->owner, *p_name), false);
+		Variant value = LuaFunction::invoke_lua(_get->method, Array::make(p_instance->owner, *p_name), false);
 		if (value != Variant()) {
 			*p_value = value;
 			return true;
@@ -121,7 +123,7 @@ GDExtensionScriptInstanceGetClassCategory get_class_category_func;
 
 GDExtensionBool property_can_revert_func(LuaScriptInstance *p_instance, const StringName *p_name) {
 	if (const LuaScriptMethod *method = p_instance->script->get_metadata().methods.getptr("_property_can_revert")) {
-		Variant result = LuaFunction::invokev_lua(method->method, Array::make(p_instance->owner, *p_name), false);
+		Variant result = LuaFunction::invoke_lua(method->method, Array::make(p_instance->owner, *p_name), false);
 		if (result) {
 			return true;
 		}
@@ -132,7 +134,7 @@ GDExtensionBool property_can_revert_func(LuaScriptInstance *p_instance, const St
 
 GDExtensionBool property_get_revert_func(LuaScriptInstance *p_instance, const StringName *p_name, Variant *r_ret) {
 	if (const LuaScriptMethod *method = p_instance->script->get_metadata().methods.getptr("_property_get_revert")) {
-		Variant result = LuaFunction::invokev_lua(method->method, Array::make(p_instance->owner, *p_name), true);
+		Variant result = LuaFunction::invoke_lua(method->method, Array::make(p_instance->owner, *p_name), true);
 		if (LuaError *error = Object::cast_to<LuaError>(result)) {
 			ERR_PRINT(error->get_message());
 		}
@@ -175,7 +177,7 @@ GDExtensionBool validate_property_func(LuaScriptInstance *p_instance, GDExtensio
 	if (const LuaScriptMethod *_validate_property = p_instance->script->get_metadata().methods.getptr("_validate_property")) {
 		PropertyInfo property_info(p_property);
 		Dictionary property_info_dict = property_info;
-		LuaFunction::invokev_lua(_validate_property->method, Array::make(p_instance->owner, property_info_dict), false);
+		LuaFunction::invoke_lua(_validate_property->method, Array::make(p_instance->owner, property_info_dict), false);
 		return true;
 	}
 	else {
@@ -196,7 +198,7 @@ GDExtensionInt get_method_argument_count_func(LuaScriptInstance *p_instance, con
 void call_func(LuaScriptInstance *p_instance, const StringName *p_method, const Variant **p_args, GDExtensionInt p_argument_count, Variant *r_return, GDExtensionCallError *r_error) {
 	if (const LuaScriptMethod *method = p_instance->script->get_metadata().methods.getptr(*p_method)) {
 		r_error->error = GDEXTENSION_CALL_OK;
-		*r_return = LuaFunction::invoke_method_lua(method->method, p_instance->owner, p_args, p_argument_count, false);
+		*r_return = LuaCoroutine::invoke_lua(method->method, VariantArguments(p_instance->owner, p_args, p_argument_count), false);
 	}
 	else {
 		r_error->error = GDEXTENSION_CALL_ERROR_INVALID_METHOD;
@@ -205,13 +207,13 @@ void call_func(LuaScriptInstance *p_instance, const StringName *p_method, const 
 
 void notification_func(LuaScriptInstance *p_instance, int32_t p_what, GDExtensionBool p_reversed) {
 	if (const LuaScriptMethod *_notification = p_instance->script->get_metadata().methods.getptr("_notification")) {
-		LuaFunction::invokev_lua(_notification->method, Array::make(p_instance->owner, p_what, p_reversed), false);
+		LuaCoroutine::invoke_lua(_notification->method, Array::make(p_instance->owner, p_what, p_reversed), false);
 	}
 }
 
 void to_string_func(LuaScriptInstance *p_instance, GDExtensionBool *r_is_valid, String *r_out) {
 	if (const LuaScriptMethod *_to_string = p_instance->script->get_metadata().methods.getptr("_to_string")) {
-		Variant result = LuaFunction::invokev_lua(_to_string->method, Array::make(p_instance->owner), false);
+		Variant result = LuaFunction::invoke_lua(_to_string->method, Array::make(p_instance->owner), false);
 		if (result) {
 			*r_out = result;
 			*r_is_valid = true;