+++ src/utils/LuaCoroutinePool.hpp
@@ -19,38 +19,23 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-#ifndef __LUA_FUNCTION_HPP__
-#define __LUA_FUNCTION_HPP__
+#ifndef __UTILS_COROUTINE_POOL_HPP__
+#define __UTILS_COROUTINE_POOL_HPP__
 
-#include "LuaObject.hpp"
-
-using namespace godot;
+#include "sol/sol.hpp"
 
 namespace luagdextension {
 
-class VariantArguments;
+struct LuaCoroutinePool {
+	LuaCoroutinePool(lua_State *L);
 
-class LuaFunction : public LuaObjectSubclass<sol::protected_function> {
-	GDCLASS(LuaFunction, LuaObject);
+	sol::thread acquire(const sol::function& f);
+	void release(const sol::thread& coroutine);
 
-public:
-	LuaFunction();
-	LuaFunction(sol::protected_function&& function);
-	LuaFunction(const sol::protected_function& function);
-
-	Variant invokev(const Array& args);
-	Variant invoke(const Variant **args, GDExtensionInt arg_count, GDExtensionCallError &error);
-
-	static Variant invoke_lua(const sol::protected_function& f, const VariantArguments& args, bool return_lua_error);
-
-	Callable to_callable() const;
-
-	const sol::protected_function& get_function() const;
-
-protected:
-	static void _bind_methods();
+private:
+	sol::state_view L;
 };
 
 }
 
-#endif  // __LUA_FUNCTION_HPP__
+#endif  // __UTILS_COROUTINE_POOL_HPP__